use stdint compliant ints; include stdint.h if possible, else fallback to glib

+++ src/rvoice/fluid_phase.h
@@ -47,7 +47,7 @@
 * It is a 64 bit number. The higher 32 bits contain the 'index' (number of
 * the current sample), the lower 32 bits the fractional part.
 */
-typedef unsigned long long fluid_phase_t;
+typedef uint64_t fluid_phase_t;
 
 /* Purpose:
  * Set a to b.
@@ -56,26 +56,26 @@ typedef unsigned long long fluid_phase_t;
  */
 #define fluid_phase_set(a,b) a=b;
 
-#define fluid_phase_set_int(a, b)    ((a) = ((unsigned long long)(b)) << 32)
+#define fluid_phase_set_int(a, b)    ((a) = ((uint64_t)(b)) << 32)
 
 /* Purpose:
  * Sets the phase a to a phase increment given in b.
  * For example, assume b is 0.9. After setting a to it, adding a to
  * the playing pointer will advance it by 0.9 samples. */
 #define fluid_phase_set_float(a, b) \
-  (a) = (((unsigned long long)(b)) << 32) \
-  | (uint32) (((double)(b) - (int)(b)) * (double)FLUID_FRACT_MAX)
+  (a) = (((uint64_t)(b)) << 32) \
+  | (uint32_t) (((double)(b) - (int)(b)) * (double)FLUID_FRACT_MAX)
 
 /* create a fluid_phase_t from an index and a fraction value */
 #define fluid_phase_from_index_fract(index, fract) \
-  ((((unsigned long long)(index)) << 32) + (fract))
+  ((((uint64_t)(index)) << 32) + (fract))
 
 /* Purpose:
  * Return the index and the fractional part, respectively. */
 #define fluid_phase_index(_x) \
   ((unsigned int)((_x) >> 32))
 #define fluid_phase_fract(_x) \
-  ((uint32)((_x) & 0xFFFFFFFF))
+  ((uint32_t)((_x) & 0xFFFFFFFF))
 
 /* Get the phase index with fractional rounding */
 #define fluid_phase_index_round(_x) \
@@ -108,7 +108,7 @@ typedef unsigned long long fluid_phase_t;
 /* Purpose:
  * Subtract b samples from a.
  */
-#define fluid_phase_sub_int(a, b)  ((a) -= (unsigned long long)(b) << 32)
+#define fluid_phase_sub_int(a, b)  ((a) -= (uint64_t)(b) << 32)
 
 /* Purpose:
  * Creates the expression a.index++. */
